move 'Label' component to 'src/components' folder
