+++ src/components/Label/Label.scss
