+++ src/modules/forms/plugins/Cors/CorsPlugin.js
@@ -8,7 +8,7 @@ import optionsTransformer from '../../../../helpers/optionsTransformer';
 import getValues from '../../../../helpers/getValues';
 
 import Row from '../../../Layout/Row/Row';
-import Label from '../../../labels/Label';
+import Label from '../../../../components/Label/Label';
 import Input from '../../../inputs/Input';
 import Hint from '../../../../components/Hint/Hint';
 import ControlBar from '../ControlBar/ControlBar';
